MM-24506: Fix closing DM/GM from header dropdown menu (#5422)
* Fix closing DM/GM from header dropdown menu

Commit 6dee486 fixed an issue with redirecting the user to direct or
group messages when they were the last ones to be visited. But a
collateral regression was introduced: closing DMs or GMs stopped
working, as the user was returned to the closed channel immediately
after leaving it.

That commit was reverted, and then applied again in commit 9193f53,
which also introduced the new leaveDirectChannel function to correctly
handle the new issue with closing DMs or GMs. But this function was only
applied to the case where the user closed the channel through the x
button on the sidebar channel item.

But the bug remained in place when the user closed the channel using the
"Close {Direct, Group} Message" option in the channel header dropdown
menu.

This commit just fixes this second use case by calling the
leaveDirectChannel function from the right place.

* Test closing direct and group messages

Add tests checking that closing a direct or group message, either
through the x button on the sidebar channel item or through the channel
header dropdown menu, behaves as expected.

* Remove stage metadata from e2e test

Former-commit-id: 60264744fa52558569fc77a402f2241286449644