make actorId optional (#7800)
The combined system message component assembles a message from a set of system posts, but the TypeScript definition assumes that all system messages have actors. Some do not, including the opening system messages of `Town Square`, for which the join message is initiated by the server and not a user.

Former-commit-id: 619b9d131e0aa5e381045822e8d32ee3ced9348f