MM-35074: short circuit slash command hooks (#7969)
When the Confluence plugin consumes a slash command for client-side processing (i.e. `/confluence subscribe`), it signals this by returning an empty object as per the required interface.

However, the webapp passes this empty object through all other registered slash command hooks. When the Confluence and Incident Collaboration plugins are installed (and Incident Collaboration "registers last"), Incident Collaboration receives this empty object and mistakenly initializes the `undefined` args parameter in testing whether it should process the slash command.  The final result consumed by the webapp is now partially initialized, leading to a JavaScript exception and a create post component that won't process slash commands until it mounts afresh.

While we can address this in Incident Collaboration, it's unexpected that the webapp would notify plugins at all in this case, given there is literally no metadata on which to try to do anything meaningful. Short-circuit in this case, avoiding the original report altogether.

Fixes: https://mattermost.atlassian.net/browse/MM-35074
Former-commit-id: b0f2df7a539524e6ca3c296df1b8624baab80b8d