MM-40317: fixes showing new replies banner (#9646)
* MM-40317: fixes showing new replies banner

Currently we rely on the new messages line visibility on whether to show
or not the "New Replies" banner. However new messages line is not showing,
by design, on unfollowed threads. Which made the new replies banner to
show when new replies arrived for those threads.

This commit fixes the issue by guarding on the list's 'visibleStopIndex',
if visibleStopIndex is zero then the last item of the list is visible.
so we show the new replies only if the visibleStopIndex is not zero.

* Removes unneeded change

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>
Former-commit-id: 1e63c8aa75247e47da2aa665f02e304dc617622d