e2e: fix 'handle new post' spec read/unread assertions (#9715)
* e2e: fix 'handle new post' spec

As far as I can tell, the assertions in these tests didn't match the
expected behavior. The first test asserted that the test user will see
a channel they were just added to as read, when they actually see
both the unread highlight and mention badge from the join system message.

There was also a problem with an implicit dependency between the first and second
tests. The second test again expected that the test channel starts as read from
the test user's point of view, but it reuses the same channel from the first test,
which ends on unread.

* reference zephyr cases for tests

Former-commit-id: d165fe4e8fa0f474545a8be7c2bcaf2b5b3ed86b