MM-43194: remove plugin-related dispatch on enable (#10152)
* MM-43194: remove plugin-related dispatch on enable

In https://github.com/mattermost/mattermost-webapp/pull/8545, a change to dispatch `RECEIVED_WEBAPP_PLUGIN` was introduced on receipt of a websocket event for enabling a plugin. Unfortunately, this had the unexpected effect of changing the semantics of plugin reloading.

In particular, for Playbooks, if that version of the plugin was already installed, instead of replacing it, it would appear to be removed altogether. The plugin registration and management is, unfortunately, very brittle.

No test is included with this PR at this time.

* revert unit test changes as well

Former-commit-id: 3e484a2a9d845390a00ca1309916b8bd069fdb68