MM-44216: Move channel header icons to App Bar (#10311)
* Hide channel header buttons if App Bar is shown

* Render channel header components in the App Bar

When the App Bar is shown, the buttons in the channel header are no
longer rendered there, so they need to be rendered in the App Bar
instead. This commit does that, and takes care of an explicit order in
the icons, so that Playbooks and Boards icons are the first ones, then
the explicitly registered App Bar icons, and finally the channel header
icons.

* Handle channel header icons in AppBarPluginComponent

The AppBarPluginComponent can now receive a channel header component,
apart from the explicit App Bar components. To handle that case, we
simply need to render its original component.icon, which is styled to
center it in the circle that encloses all icons in the App Bar.

* Update ChannelHeaderPlug snapshots

* Order the app bar components instead of filtering

* Add snapshot test when shouldShowAppBar is true

* Use scss file instead of styled-components

* Update AppBar snapshot

* MM-44221: Add admin console setting to enable/disable App Bar (#10318)

* Add admin console config for App Bar

* make i18n-extract

* Fix current tests

* Test null rendering of App Bar when it is disabled

* Hide the AppBar admin setting in Cloud

* Hide channel header buttons if App Bar is shown

* Render channel header components in the App Bar

When the App Bar is shown, the buttons in the channel header are no
longer rendered there, so they need to be rendered in the App Bar
instead. This commit does that, and takes care of an explicit order in
the icons, so that Playbooks and Boards icons are the first ones, then
the explicitly registered App Bar icons, and finally the channel header
icons.

* Handle channel header icons in AppBarPluginComponent

The AppBarPluginComponent can now receive a channel header component,
apart from the explicit App Bar components. To handle that case, we
simply need to render its original component.icon, which is styled to
center it in the circle that encloses all icons in the App Bar.

* Update ChannelHeaderPlug snapshots

* Order the app bar components instead of filtering

* Add snapshot test when shouldShowAppBar is true

* Use scss file instead of styled-components

* Update AppBar snapshot

* Address review help text fix

* Fix order of plugins registering an App Bar icon

We need to push every plugin under Playbooks and Boards, so we need to
return 1 instead of 0 when the plugin ID is not the one from Playbooks
or Boards.

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>

* Fix colors in channel header icons

* Update snapshot with npm test -- -u app_bar

* App bar shown with plugins without app bar icons

When there are only plugins with only channel header icons registered,
the app bar should still be shown.

Co-authored-by: Mattermod <mattermod@users.noreply.github.com>
Former-commit-id: 84462fa68824b75a28d7d47dc8d7472f469fd15c