Extract SkeletonLoader to @mattermost/components (#11011)
Skeleton loading component is now starting to be used newer designs of product. So i thought it would be a good idea to extract the component to package '@mattermost/components' so other products can use it as well. I have tried to keep the API of the component as simple and generic as possible to be used easily in other plugins/products.

Former-commit-id: 7712149df5b1ac7622a7406a1e86d0252720106f